Update Delete Category

+++ drivers/mysql/category/category.go
@@ -59,7 +59,12 @@ func (r *CategoryRepo) UpdateCategory(id int, category *entities.Category) (*ent
 func (r *CategoryRepo) DeleteCategory(id int) error {
 	complaints := r.DB.Where("category_id = ?", id).Find(&entities.Complaint{})
 	if complaints.RowsAffected > 0 {
-		return constants.ErrCategoryHasComplaints
+		return constants.ErrCategoryHasBeenUsed
+	}
+
+	news := r.DB.Where("category_id = ?", id).Find(&entities.News{})
+	if news.RowsAffected > 0 {
+		return constants.ErrCategoryHasBeenUsed
 	}
 
 	if err := r.DB.Delete(&entities.Category{}, id).Error; err != nil {
